modif sql

+++ admin/create.sql
@@ -4,7 +4,7 @@ DROP TABLE Activity;
 DROP TABLE Day;
 
 CREATE TABLE User (
-  IdU           INT NOT NULL AUTO_INCREMENT, -- max 10 user
+  IdU           TINYINT(2) NOT NULL AUTO_INCREMENT, -- max 10 user
   NomU          VARCHAR(40),
   PrenomU       VARCHAR(40),
   LoginU        VARCHAR(40),
@@ -13,15 +13,15 @@ CREATE TABLE User (
 );
 
 CREATE TABLE Activity (
-  IdA    INT NOT NULL,
+  IdA    TINYINT NOT NULL,
   NomA   VARCHAR(40),
-  DureeA INT,
+  DureeA TINYINT(2),
   PRIMARY KEY (IdA)
 );
 
 CREATE TABLE Day (
   DateD DATE NOT NULL,
-  IdU   INT,
-  IdA   INT,
+  IdU   TINYINT(2),
+  IdA   TINYINT(2),
   PRIMARY KEY (DateD)
 );
